Change form pages files to project-developer-form
