--- frontend/containers/project-developer-form-pages/index.ts
@@ -1,5 +1,0 @@
-export { default as Profile } from './pages/profile';
-export { default as About } from './pages/about';
-export { default as ProjectDeveloperOutro } from './pages/project-developer-outro';
-
-export type { ProfileProps } from './types';
