+++ frontend/containers/project-developer-form/types.ts
@@ -1,8 +1,17 @@
+import {
+  SetFieldValue,
+  UseFormClearErrors,
+  UseFormRegister,
+  UseFormSetError,
+  FormState,
+  Control,
+} from 'react-hook-form';
 import { UseMutationResult } from 'react-query';
 
 import { AxiosResponse, AxiosError } from 'axios';
 
-import { ProjectDeveloper, ProjectDeveloperSetupForm } from 'types/projectDeveloper';
+import { Enum } from 'types/enums';
+import { ProjectDeveloper, ProjectDeveloperSetupForm, Interest } from 'types/projectDeveloper';
 
 import { ErrorResponse } from 'services/types';
 
@@ -22,4 +31,40 @@ export type ProjectDeveloperFormProps = {
     ProjectDeveloperSetupForm,
     unknown
   >;
+};
+
+export type ProfileProps = {
+  /**  React-hook-form setError function */
+  setError: UseFormSetError<ProjectDeveloperSetupForm>;
+  /**  React-hook-form setValue function */
+  setValue: SetFieldValue<any>;
+  /**  React-hook-form register function */
+  register: UseFormRegister<ProjectDeveloperSetupForm>;
+  /**  React-hook-form clearErrors function */
+  clearErrors: UseFormClearErrors<ProjectDeveloperSetupForm>;
+  /**  React-hook-form state - errors */
+  errors: FormState<ProjectDeveloperSetupForm>['errors'];
+  /**  React-hook-form control function */
+  control: Control<ProjectDeveloperSetupForm, any>;
+  /**  Project developers enums */
+  projectDeveloperTypeEnums: Enum[];
+  /**  If there is some error on enum query */
+  enumsIsError: boolean;
+  /**  PD picture small url */
+  picture?: string;
+};
+
+export type AboutProps = {
+  /** Interest enums */
+  interests: Interest[];
+  /**  If thgere is some error on enum query */
+  enumsIsError: boolean;
+  /**  React-hook-form register function */
+  register: UseFormRegister<ProjectDeveloperSetupForm>;
+  /**  React-hook-form setValue function */
+  setValue: SetFieldValue<any>;
+  /**  React-hook-form state - errors */
+  errors: FormState<ProjectDeveloperSetupForm>['errors'];
+  /**  React-hook-form clearErrors function */
+  clearErrors: UseFormClearErrors<ProjectDeveloperSetupForm>;
 };
